Auto-export Spicy-generated events inside the Zeek interpreter.
We already did this partially before but didn't cover all the cases. The
trick is that we need to install a global ID for the event so that we
can export it; however we won't know the type of the event because Spicy
lazily works with whatever handlers are defined. The code accounts for
that now and sets a dummy event type with no arguments for the ID if
there are no handlers at all.

Closes #15.