Add support for a custom Zeek scripts directory.
This is where scripts accompanying installed Zeek analyzers go. The path
can be printed through `spicyz --print-scripts-dir`, it'll be
`<libdir>/spicy/zeek/scripts`. If ZEEKPATH is not set explicitly, the
plugin will add this to Zeek's search path.

We are also relocating the module directory to
`<libdir>/spicy/zeek/modules`.

This is in preparation for https://github.com/zeek/spicy-analyzers/issues/31.