Support recursive file analysis.
So far we only supported analysis of file content coming directly out of
a protocol analyzer; we did not support passing on files to Zeek that
were nested inside another file. This missing piece was that our file
IDs weren't taking recursive nesting into account. This change reworks
that part.