Rework initializing of file meta data.
The previous change for file-in-file analysis caused a regression for
files coming out of connection analysis: not all files.log fields were
filled out anymore for them. This fixes that by ensuring that we do pass
the connection information into the Zeek-side API whereever we can. It's
actually doing a bit more that we'd really need to do there, but this
way it follows the Zeek API more closely rather than relying on some
undocumented behavior.

The reason we didn't catch the regression is that we hadn't bumped the
submodules in the Spicy repository yet, meaning the analyzers weren't
tested against the new plugin code.