Make parallel build and test opt-out instead of opt-in.
In b06e4bf741f we made the package build and test process sequential to
not overwhelm machines with too little RAM. This had the side effect of
making _all builds_ slower.

In this patch we revise that fix but going back to package build and
test parallel by default, but with an option to opt out of this. If the
environment variable `SPICY_ZKG_PROCESSES` is set, we will use that as
the number of build or test jobs to spawn; otherwise we default to a
hardcoded level of parallelism.