Fix parsing of evt files not ending in newlines.
When parsing evt files, we previously would manually and explicitly
handle EOF encountered when extracting lines from evt files. This lead
to us incorrectly ignoring the last line if it did not end in a newline.

With this patch we now use the standard pattern of extracting lines
until `getline` returns an error.

Closes #56.