Extend `spicy_add_analyzer` to allow installing additional Zeek scripts.
This patch changes `spicy_add_analyzer` to also support specifying
scripts which should be installed with the plugin. For that we change
the function to now support named arguments (`NAME`, `SOURCES`, and
`SCRIPTS`). For now we continue to support the previous format where
arguments were all unnamed, however this form does not support
installing scripts.

Cloes #60.