Switch file analysis state to stack of files currently in flight.
This commit prepares for feeding data into multiple files concurrently
by switchting the internal state from representing just a single file to
a stack of all files currently in flight.

We do not yet actually make use of the new structure; for now we
continue to support just a single file from the API perspective.
Concurrent analysis will be added in the next commit.