Support forwarding to Zeek TCP-level application data from a UDP analyzer.
This necessarily remains a bit hackish because of a lack of better
Zeek-side APIs: we have to create a fake TCP analyzer that's just
there for our actual child analyzer to have one to access; no further
TCP semantics are provided.

This comes with a new check that any connected analyzer is indeed a
TCP application-layer analyzer. We were already limited to that case,
but didn't catch it so far. We could open things up to UDP in the
future as well, but would need a way for the user to select which one
to use. I'll leave that for later once a use-case comes up.

Closes #92.