Do not reconfigure already configured CMake project.
Previously we would always configure the plugin which could have found a
`spicy-config` somewhere in the system. If we are building spicy-plugin
inside Zeek which also includes a Spicy, this cannot work as
spicy-config would not be available at configure time.

This patch makes the execution of `configure` conditional so we can
disable it and parameterize the build from the outside.