Change how parse errors are reported.
We used to write them to weird.log, which isn't a good fit. Now we do
the following:

- Parse errors during protocol and packet parsing go through the DPD
  confirmation/violation pipeline, meaning Zeek will log errors into
  `dpd.log` if they happen after a positive confirmation, otherwise
  just silently ignore them. (For packet analysis, this applies only
  for Zeek >= 4.2; for older versions we silently ignore errors
  because those don't support packet analyzers reporting them.)

- Parse errors during file parsing will be silently ignored because
  Zeek currently doesn't have a mechanism to report them. (Not great,
  obviously.)

In all cases, we track parse errors inside the debug output so that
there's a way to confirm what's happening.

We now also implicitly raise a protocol violation for any parse errors
that we couldn't recover from.

Closes #101.