Make names of generated hooks deterministic.
When generating hook names we previously used the address of the
specific event instances as a way to make names unique. Due to that the
generated names were unlikely to be the same for different invocations
of `spicyz`.

With this patch we add basic hashing for `glue::Event`s so that we can
deterministically generate unique identifiers for events. This should
speed up JIT whenever EVT files were used.