Readd support for `protocol_confirmation`.
With zeek-5.1 the deprecated `protocol_confirmation` and
`procotol_violation` events will be removed. Until then whenever they
are triggered, Zeek will also trigger the new events
`analyzer_confirmation` and `analyzer_violation`. Since the new events
are emitted for more than just classic protocol analyzers, Zeek will not
invoke handlers for the old events if the new events are triggered.

When we adjusted the code to used to emit confirmation and violation
events in f8fde8d we completely moved to the new infrastructure. This
meant that even if the plugin was running against zeek-4.2 existing
handlers for `protocol_confirmation` and `procotol_violation` in user
code were not called anymore which unexpectedly forced the zeek-5.1 API
on users.

This regression is addressed in this patch. We now continue to use the
old events until zeek-5.1 which will ensure that user code handlers
forboth old as well as the new events gets invoked. While usually Zeek
ensures that `protocol_confirmation` or `analyzer_confirmation` is
called exactly once, it will still invoke both `protocol_confirmation`
and `analyzer_confirmation` if both are implemented.

Closes #112.