Prevent passing an invalid session when reporting analyzer violation.
The session of a packet might be unset while the Zeek API implicitly
requires a valid session when using Analyzer::AnalzerViolation`. Since
we previously did not check that the session was valid, we might have
sent invalid data into Zeek which would it have caused to segfault.

With this patch we only ever report a violation via the Zeek API if the
packet had a valid session.

Closes #132.