Fix spelling of Zeek-side tags for packet analyzers.
They used to turn out as `ANALYZER_SPICY__FOO`. Now they are
`ANALYZER_SPICY_FOO`.

We also unify how the derivation of the Zeek-side name happens across
all three types of analyzers. On the Spicy side, analyzer names now
remain in their original spelling (i.e., `Spicy::Foo` instead of
`Spicy_Foo`). On the Zeek-side, the name for packet analyzers is now
`Spicy_Foo` instead of `Spicy::Foo` (that was the case for
protocol/file analyzers already).

On the Zeek side, the removal of the double-underscore and the name
change for packet analyzer is user-visible and could break some
scripts. It'll most likely be encountered in logs and, more
importantly, in scripts configuring a packet analyzer through calls to
`register_packet_analyzer_by_name()` and
`try_register_packet_analyzer_by_name()`. Calls to the former will
need to have double-underscores removed, and calls to the latter will
need to switch from `Spicy::Foo` to `Spicy_Foo`. This is a bit
unfortunate, but I don't see a way to avoid it if we want to become
consistent; and doing that change now seems a better time than later.