Record documentation strings for events in EVT files.
Event definitions can now be prefixed with Zeek-style documentation
strings (`##`), which will be extracted and retaining inside the
internal event data structure for later use.

TODO: This hasn't been exercised; we don't even have way to output the
results for testing yet. Leaving it in a branch for the time being in
case we want to come back to it.