Adjust plugin for changes for zeek/spicy-1326.
To address zeek/spicy-1326 we changed the code generation for enums to
emit a proper runtime type, in particular a type which has defined
semantics for value initialization.

This patch adds compatibility code so the plugin can deal with these
generated enums types in addition to the previous format for backwards
compatibility.