Push includes from `zeek-compat.h` into actual users.
Now that `zeek-compat.h` provides only minimal compatibility wrappers it
makes sense to push includes into users. We also remove a hack which
caused Zeek header to be included with `DEBUG` in Zeek debug builds;
this was never true for all headers (direct includes from Zeek
elsewhere), and also was not 100% clean in that it would cause us to
potentially use different implementations of the same function (once
with `DEBUG` and once without).

With this change `zeek-compat.h` only exports version macros and a
backport of a rename from `bro_int_t` to `zeek_int_t`.