Revert moving Zeek includes out of `zeek-compat.h.`
This reverts most of `5313ab `. The problem is that the Zeek
includes need some consistent handling in terms of preprocessor setup
(`DEBUG`) and inclusion of `zeek-config.h`. The old `zeek-compat.h`
used to handle that, but we lost it by moving the includes over into
individual files. As a result, I saw some ill-defined crashes inside
the plugin, depending on local configuration/environment.

This reverts most of the previous the change because it's easier/safer
to handle this centrally than adding the logic too all files needing
to include some Zeek header.