Adapt Spicy's changes to exception hierarchy.
We now consistently report runtime errors (including actual parsing
errors) as analyzer violations; and o longer as analyzer errors. The
former go into DPD and `analyzer.log`, while the latter would end up
in `reporter.log` (which isn't really helpful).

The commit also takes the opportunity to report file analyzer
violations through Zeek's new API for that when available; and passes
offending data along with the violation for logging on the Zeek side.

This goes with https://github.com/zeek/spicy/pull/1362. It remains
working with older Spicy versions, too, but for them the output may
change (which is why we disable one test for older version). With that
branch in use, the output is what we want.

Closes #160.