Add explicit dependency of spicyz on its runtime files.
If spicy-plugin is built as part of Zeek we want to be able to use a
simple dependency on `spicy` when declaring Spicy analyzers in the Zeek
tree (this is in fact exactly what the macro `spicy_add_analyzer` does).
In order for `spicy` to work when built into Zeek we need its runtime
dependencies like e.g., `zeek_rt.hlt` in place before using it.

This patch adds an explicit dependency of `spicyz` on
`copy-zeek-spicy-dist-files` which achieves just that.