Avoid repeated redundant calls to `globals` in generated code.
Previously we would perform repeated, identical calls to `globals` in
hook bodies generated for interfacing with Zeek event handlers in order
to look up the generated event handler to the point where this became
noticable in perf traces.

Since the event handler does not change during the hook body, we now
cache the globals lookup in a local variable.