Fix backport of change to how exceptions are reported.
47e1e95cb149fb48a4a56473dc51ffeadcc1442b changed how parse errors are
reported to Zeek. In backporting this change we missed that the changes
depended on the Spicy-side exception, namely that a `RecoverableFailure`
was a `ParseError`. This is not the case for older Spicy versions.

This patch fixes the backport so we handle `RecoverableFailure`
exceptions as well. This prevents exception escape from that.

Closes #195.